move io and resize functions into separate modules
