+++ pystiche/image/transforms/functional/__init__.py
@@ -1,3 +1,5 @@
-from .misc import *
+from .io import *
+from .resizing import *
 from .motif import *
 from .color import *
+from .misc import *
